HCK-14286: Add logic to alter select statement
