+++ forward_engineering/alterScript/alterScriptHelpers/viewHelpers/alterNameHelper.js
@@ -1,6 +1,7 @@
 const { AlterScriptDto } = require('../../types/AlterScriptDto');
-const { getFullViewName, getSchemaOfAlterView } = require('../../../utils/general');
+const { getFullCollectionName, getSchemaOfAlterView } = require('../../../utils/general');
 const { getKeys } = require('./getKeys');
+const { createView } = require('./createView');
 
 /**
  * @param {Object} view
@@ -22,23 +23,10 @@ const getRenameViewScriptDto = (view, ddlProvider, mapProperties) => {
 		return undefined;
 	}
 
-	const oldFullViewName = getFullViewName({ ...viewSchema, code: oldName, name: oldName });
+	const oldFullViewName = getFullCollectionName({ ...viewSchema, code: oldName, name: oldName });
 	const dropScript = ddlProvider.dropView({ viewName: oldFullViewName });
 
-	const schemaData = { schemaName: viewSchema.schemaName || '' };
-	const viewData = {
-		name: newName,
-		keys: getKeys({
-			viewSchema,
-			ddlProvider,
-			mapProperties,
-			collectionRefsDefinitionsMap: view.compMod?.collectionData?.collectionRefsDefinitionsMap ?? {},
-		}),
-		schemaData,
-	};
-
-	const hydratedView = ddlProvider.hydrateView({ viewData, entityData: [viewSchema] });
-	const createScript = ddlProvider.createView(hydratedView, {}, viewSchema.isActivated);
+	const createScript = createView(ddlProvider, mapProperties, view);
 
 	return AlterScriptDto.getInstance([dropScript, createScript], true, false);
 };
