+++ forward_engineering/ddlProvider/templates.js
@@ -64,6 +64,4 @@ module.exports = {
 	renameColumn: 'ALTER TABLE ${tableName} RENAME COLUMN ${oldColumnName} TO ${newColumnName};',
 
 	dropView: 'DROP VIEW ${viewName};',
-
-	renameView: 'RENAME VIEW ${oldViewName} TO ${newViewName};',
 };
